backend/api/gantt/requirements.txt
# SwissAI Gantt Planner API - Dependencies
# Self-contained requirements for the Gantt planning API module

# Core dependencies
requests>=2.28.0        # For HTTP API calls to Public AI
python-dotenv>=1.0.0    # Environment variable management
pydantic>=2.0.0         # Data validation and schema definition

# FastAPI dependencies  
fastapi>=0.104.0        # Web framework
uvicorn>=0.24.0         # ASGI server

# Optional dependencies for enhanced functionality
python-multipart>=0.0.6 # For form data support
httpx>=0.24.0           # Alternative HTTP client (optional)